% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{add_predteller}
\alias{add_predteller}
\title{add_predteller}
\usage{
add_predteller(dt, parameters, globs)
}
\arguments{
\item{parameters}{cube parameters}

\item{globs}{global parameters}

\item{TNF}{merged teller-nevner file}
}
\description{
Adds predicted teller for age- and gender standardization

PREDRATE
Maa JUKSE DET TIL LITT MED NEVNER 0. Bruken her er jo slik at dette er tomme celler,
og ikke minst vil raten nesten garantert skulle ganges med et PREDTELLER=0
Tillater TELLER<=2 for aa unngaa evt numeriske problemer. Virker helt uskyldig gitt bruken
}
