% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{aggregate_to_periods}
\alias{aggregate_to_periods}
\title{aggregate_to_moving_average}
\usage{
aggregate_to_periods(dt, parameters)
}
\arguments{
\item{dt}{KUBE to be aggregated}

\item{parameters}{cube parameters}

\item{reset_rate}{Reset RATE after aggregating to periods?}
}
\description{
Finn "snitt" for ma-aar.
DVs, egentlig lages forloepig bare summer, snitt settes etter prikking under
Snitt tolerer missing av type .f=1 ("random"), men bare noen faa anonyme .f>1, se KHaggreger
Rapporterer variabelspesifikk VAL.n som angir antall aar brukt i summen naar NA holdt utenom

Dersom is_movav = FALSE, legges val.n til for alle verdikolonner, satt til 1 dersom originale snitt
og satt til antall år dersom originale summer.
}
