% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{AggregerRader}
\alias{AggregerRader}
\title{AggregerRader (kb)}
\usage{
AggregerRader(FG, nyeexpr, FGP)
}
\arguments{
\item{FG}{}

\item{nyeexpr}{}

\item{FGP}{}
}
\description{
AggregerRader (kb)
}
