% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoring.R
\name{AnonymiserNaboer}
\alias{AnonymiserNaboer}
\title{AnonymiserNaboer (kb)}
\usage{
AnonymiserNaboer(
  FG,
  ovkatstr,
  FGP = list(amin = 0, amax = 120),
  ref_year_type = "Moving",
  globs = SettGlobs()
)
}
\arguments{
\item{FG}{}

\item{ovkatstr}{}

\item{FGP}{}

\item{ref_year_type}{}
}
\description{
AnonymiserNaboer (kb)
}
