% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHfilgruppe.R
\name{check_encoding}
\alias{check_encoding}
\title{check_encoding}
\usage{
check_encoding(dt)
}
\arguments{
\item{dt}{file group}
}
\value{
a list of unique problematic values which indicates that a specific file should be read with different encoding
}
\description{
Scans all character columns for potential encoding issues.
Searches for \verb{<c3><a6>}, etc., which indicates UTF-8 read by a single-byte locale.
Searches for \code{Ã}, which indicates UTF-8 bytes were misinterpreted as Latin-1 characters.
}
