% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkube.R
\name{check_if_lagkube_available}
\alias{check_if_lagkube_available}
\title{check_if_lagkube_available}
\usage{
check_if_lagkube_available()
}
\description{
Checks if guardfile exists, indicating that the system is already running.
If file doesn't exist, or if user overrides and force continue, the file
is generated and TRUE is returned indicating that the function may continue.
If the file exists and the user does not override, FALSE is returned indicating
that the system is busy and data processing stops.

An on.exit call must be included in the main function to delete
the file when the function finish or crash. This function checks if the file already exists,
and generate the file if not (or overridden by user).
}
