% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{do_aggregate_file}
\alias{do_aggregate_file}
\title{do_aggregate_file}
\usage{
do_aggregate_file(file, valsumbardef = list(), globs = SettGlobs())
}
\arguments{
\item{file}{}

\item{valsumbardef}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
do_aggregate_file
}
