% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{do_balance_missing_teller_nevner}
\alias{do_balance_missing_teller_nevner}
\title{do_balance_missing_teller_nevner}
\usage{
do_balance_missing_teller_nevner(dt)
}
\arguments{
\item{dt}{data}
}
\description{
Maa "balansere" NA i teller og nevner slik sumrate og sumnevner balanserer.
Kunne med god grunn satt SPVFLAGG her og saa bare operert med denne som en egenskap for hele linja i det som kommer
Men for aa ha muligheten for aa haandtere de forskjellige variablene ulikt og i full detalj lar jeg det staa mer generelt
Slik at dataflyten stoetter en slik endring

Om enkeltobservasjoner ikke skal brukes, men samtidig tas ut av alle summeringer
kan man ha satt VAL=0,VAL.f=-1
Dette vil ikke oedelegge summer der tallet inngaar. Tallet selv, eller sumemr av kun slike tall, settes naa til NA
Dette brukes f.eks naar SVANGERROYK ekskluderer Oslo, Akershus. Dette er skjuling, saa VAL.f=3
}
