% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{do_filter_periods_with_missing_original}
\alias{do_filter_periods_with_missing_original}
\title{filter_periods_with_missing_original}
\usage{
do_filter_periods_with_missing_original(dt)
}
\arguments{
\item{dt}{data}
}
\description{
Filters out averages with too much missing from original data. The
tolerance is set by the option anon_tot_tol in the config file, and
for rows where the proportion of original val.f == 3 (as indicated
by the helper column val.fn3) > the tolerance, the average is set to
NA and val.f = 3
}
