% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{do_handle_indata_periods}
\alias{do_handle_indata_periods}
\title{handle_indata_periods}
\usage{
do_handle_indata_periods(dt, parameters)
}
\description{
If indata contains periods, val.n must be adjusted accordingly
}
