% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{do_redesign_recode_file}
\alias{do_redesign_recode_file}
\title{do_redesign_recode_file}
\usage{
do_redesign_recode_file(filename, filedesign, tndesign, globs)
}
\arguments{
\item{filename}{}

\item{filedesign}{}

\item{tndesign}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
do_redesign_recode_file
}
