% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_filegroup_values.R
\name{do_scale_val}
\alias{do_scale_val}
\title{do_scale_val}
\usage{
do_scale_val(dt, val, parameters)
}
\description{
Scales value-columns based on information in "SKALA_VALX"-columns in ACCESS
}
