% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoring.R
\name{do_stata_censoring}
\alias{do_stata_censoring}
\title{do_stata_censoring}
\usage{
do_stata_censoring(dt, spc, batchdate, stata_exe)
}
\arguments{
\item{dt}{data file to be censored}

\item{spc}{KUBE spec}

\item{batchdate}{batchdate, to be used for temporary files}

\item{stata_exe}{path to STATA program}
}
\description{
Reads censoring information from ACCESS.
If Stata_PRIKK parameters are set, the STATA censoring script (by Jørgen Meisfjord)
is run using the do_stata_processing function
}
