% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{do_write_parquet_duckdb}
\alias{do_write_parquet_duckdb}
\title{do_write_parquet_duckdb}
\usage{
do_write_parquet_duckdb(con, tablename, filepath)
}
\arguments{
\item{con}{duckdb connection}

\item{tablename}{tabellen som skal skrives}

\item{filepath}{sti til der filen skal skrives til}
}
\description{
Skriver parquet-fil fra tabell i duckdb
}
