% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{.SetKubeParameters}
\alias{.SetKubeParameters}
\title{.SetKubeParameters (VL)}
\usage{
.SetKubeParameters(KUBEid)
}
\arguments{
\item{KUBEid}{name of kube to test on}
}
\description{
for testing LagKUBE, store all the parameters to global env
}
