% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{EkstraherRadSummer}
\alias{EkstraherRadSummer}
\title{EkstraherRadSummer (kb)}
\usage{
EkstraherRadSummer(
  FIL,
  pstrorg,
  FGP = list(amin = 0, amax = 120),
  globs = SettGlobs()
)
}
\arguments{
\item{FIL}{}

\item{pstrorg}{}

\item{FGP}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
EkstraherRadSummer (kb)
}
