% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{find_design_after_filter}
\alias{find_design_after_filter}
\title{find_design_after_filter}
\usage{
find_design_after_filter(
  file,
  parameterlist,
  originaldesign = NULL,
  outpredfilter = TRUE,
  globs = SettGlobs()
)
}
\arguments{
\item{file}{name of file to design}

\item{parameterlist}{parameters}

\item{originaldesign}{original design if existing. Default is design for file.}

\item{outpredfilter}{shuld columns in predfilter be removed, default = TRUE}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
find_design_after_filter
}
