% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{find_dims_for_stataprikk}
\alias{find_dims_for_stataprikk}
\title{find_dims (vl)}
\usage{
find_dims_for_stataprikk(dt, etabs)
}
\description{
Helper function for kube_spec, finding dimensions in KUBE
}
