% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_filedesign.R
\name{find_filedesign}
\alias{find_filedesign}
\title{find_filedesign}
\usage{
find_filedesign(file, fileparameters = NULL, globs = SettGlobs())
}
\arguments{
\item{file}{filegroup}

\item{fileparameters}{parameters from ACCESS}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
Finds design parameters for a file
}
