% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_redesign.R
\name{find_redesign}
\alias{find_redesign}
\title{find_redesign}
\usage{
find_redesign(
  fradesign,
  tildesign,
  SkalAggregeresOpp = character(),
  globs = SettGlobs()
)
}
\arguments{
\item{fradesign}{}

\item{tildesign}{}

\item{SkalAggregeresOpp}{skal noen deler evt aggregeres opp?}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
find_redesign
}
