% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{FinnDesign}
\alias{FinnDesign}
\title{FinnDesign (kb)}
\usage{
FinnDesign(FIL, FGP = list(amin = 0, amax = 120), parameters)
}
\arguments{
\item{FIL}{}

\item{FGP}{}

\item{parameters}{global parameters}
}
\description{
Brukes i lagkube og lagfilgruppe
}
