% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{FinnFellesTab}
\alias{FinnFellesTab}
\title{FinnFellesTab (kb)}
\usage{
FinnFellesTab(DF1, DF2, parameters)
}
\arguments{
\item{parameters}{global parameters}
}
\description{
FinnFellesTab (kb)
}
