% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{FinnFil}
\alias{FinnFil}
\title{FinnFil (kb)}
\usage{
FinnFil(
  FILID,
  versjonert = FALSE,
  batch = NA,
  ROLLE = "",
  TYP = "STABLAORG",
  IDKOLS = FALSE,
  globs = SettGlobs()
)
}
\arguments{
\item{FILID}{}

\item{versjonert}{}

\item{batch}{}

\item{ROLLE}{}

\item{TYP}{}

\item{IDKOLS}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
FinnFil (kb)
}
