% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{FinnKodebokForEtIntervall}
\alias{FinnKodebokForEtIntervall}
\title{FinnKodebokForEtIntervall (kb)}
\usage{
FinnKodebokForEtIntervall(Find, FRA, TILint, OVLP, letn, result, utcolnavn)
}
\arguments{
\item{Find}{}

\item{FRA}{}

\item{TILint}{}

\item{OVLP}{}

\item{letn}{}

\item{result}{}

\item{utcolnavn}{}
}
\description{
FinnKodebokForEtIntervall (kb)
}
