% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{FinnKodebokIntervaller}
\alias{FinnKodebokIntervaller}
\title{FinnKodebokIntervaller (kb)}
\usage{
FinnKodebokIntervaller(FRA, TIL, delnavn = "INT")
}
\arguments{
\item{FRA}{}

\item{TIL}{}

\item{delnavn}{}
}
\description{
FinnKodebokIntervaller (kb)
}
