% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{FinnKubeDesign}
\alias{FinnKubeDesign}
\title{FinnKubeDesign (kb)}
\usage{
FinnKubeDesign(
  KUBEdscr,
  ORGd,
  bruk0 = TRUE,
  FGP = list(amin = 0, amax = 120),
  globs = SettGlobs()
)
}
\arguments{
\item{KUBEdscr}{}

\item{ORGd}{}

\item{bruk0}{}

\item{FGP}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
FinnKubeDesign (kb)
}
