% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{FinnKubeDesignB}
\alias{FinnKubeDesignB}
\title{FinnKubeDesignB (kb)}
\usage{
FinnKubeDesignB(
  KUBEdscr,
  ORGd,
  FGP = list(amin = 0, amax = 120),
  globs = SettGlobs()
)
}
\arguments{
\item{ORGd}{}

\item{FGP}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
FinnKubeDesignB (kb)
}
