% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{FinnRedesign}
\alias{FinnRedesign}
\title{FinnRedesign (kb)}
\usage{
FinnRedesign(fradesign, tildesign, SkalAggregeresOpp = character(), parameters)
}
\arguments{
\item{fradesign}{}

\item{tildesign}{}

\item{SkalAggregeresOpp}{skal noen deler evt aggregeres opp?}

\item{parameters}{global parameters}
}
\description{
FinnRedesign (kb)
}
