% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{FinnRedesignForFilter}
\alias{FinnRedesignForFilter}
\title{FinnRedesignForFilter (kb)}
\usage{
FinnRedesignForFilter(ORGd, Filter, globs = SettGlobs())
}
\arguments{
\item{ORGd}{}

\item{Filter}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
FinnRedesignForFilter (kb)
}
