% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_original_file.R
\name{format_innlesarg_as_list}
\alias{format_innlesarg_as_list}
\title{format_innlesarg_as_list}
\usage{
format_innlesarg_as_list(args)
}
\arguments{
\item{args}{INNLESARG}
}
\value{
a list of argument name-value pairs, or empty list if no INNLESARG provided
}
\description{
Splitter args ved komma, beholder uttrykk i hermetegn eller c()
}
