% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friskvik.R
\name{generate_and_export_all_friskvik_indicators}
\alias{generate_and_export_all_friskvik_indicators}
\title{generate_and_export_all_friskvik_indicators}
\usage{
generate_and_export_all_friskvik_indicators(dt, parameters)
}
\arguments{
\item{dt}{ALLVIS cube}

\item{parameters}{global parameters}
}
\description{
Loops over requested friskvik files and export the file as csv
}
