% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friskvik.R
\name{generate_friskvik_indicator}
\alias{generate_friskvik_indicator}
\title{generate_friskvik_indicator}
\usage{
generate_friskvik_indicator(dt, id, parameters)
}
\arguments{
\item{dt}{ALLVIS cube}

\item{id}{friskvik ID}

\item{parameters}{parameters}
}
\description{
generate_friskvik_indicator
}
