% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friskvik.R
\name{generate_specific_friskvik_indicators}
\alias{generate_specific_friskvik_indicators}
\title{generate_specific_friskvik_indicators}
\usage{
generate_specific_friskvik_indicators(
  cubename = NULL,
  friskvik_id = NULL,
  year = getOption("khfunctions.year")
)
}
\arguments{
\item{cubename}{navn}

\item{friskvik_id}{Kan evt spesifisere FRISVIK_ID. Dersom denne er NULL lages alle}

\item{year}{friskvikår, default = getOption("khfunctions.year")}
}
\description{
Lager friskvikfiler fra allerede godkjent kube
}
