% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeparameters.R
\name{get_cubeparameters}
\alias{get_cubeparameters}
\title{get_cubeparameters}
\usage{
get_cubeparameters(user_args = list())
}
\arguments{
\item{args}{arguments passed to LagKUBE}
}
\value{
A list of relevant parameters
}
\description{
Helper function for \code{LagKUBE()}.
The function reads tables from ACCESS to get information on which files are used and how they need to be handled.
}
