% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgparameters.R
\name{get_filegroup_parameters}
\alias{get_filegroup_parameters}
\title{get_filegroup_parameters}
\usage{
get_filegroup_parameters(name, versjonert, dumps)
}
\arguments{
\item{filegroup_name}{name of filgruppe}

\item{batchdate}{date of current batch. Needed to filter out active parameters from ACCESS}

\item{globs}{global parameters, defaults to SettGlobs}
}
\value{
A list of relevant parameters
}
\description{
Helper function for \code{LagFilgruppe()}.
The function reads tables from ACCESS to get information on which files are used and how they need to be handled.
}
