% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeparameters.R
\name{get_filenames}
\alias{get_filenames}
\title{get_filenames}
\usage{
get_filenames(parameters, isrefverdiP, isSTNP = F)
}
\arguments{
\item{parameters}{parameters list, to access TNP and STNP with information on which files are needed}

\item{isrefverdiP}{TRUE/FALSE, is REFVERDI_VP column in KUBER table == "P", indicating if standard files should be included}

\item{isSTNP}{TRUE/FALSE, indicating if standard files should be different from the main files. This is indicated in the STANDARDTNFIL in TNP_PROD table.}
}
\description{
Helper function for \code{\link[=get_cubeparameters]{get_cubeparameters()}}
Gets a list of files specifies in the TNP_PROD table (TELLERFIL, NEVNERFIL etc.)
}
