% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeparameters.R
\name{get_filfiltre}
\alias{get_filfiltre}
\title{get_filfiltre}
\usage{
get_filfiltre(files = NULL, validdates, globs)
}
\arguments{
\item{files}{list of files generated by \code{\link[=get_filenames]{get_filenames()}}}

\item{validdates}{valid dates to specify active rows in the ACCESS table}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
Helper function for \code{\link[=get_cubeparameters]{get_cubeparameters()}}
Reads from table FILFILTRE in the ACCESS database, to get information on special treatment of files
}
