% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHglobs.R
\name{get_global_parameters}
\alias{get_global_parameters}
\title{get_global_parameters}
\usage{
get_global_parameters()
}
\value{
list
}
\description{
Sets global parameters used for filgruppe and kube. The two main functions will add specific parameters needed.
}
