% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godkjentmappe.R
\name{godkjent}
\alias{godkjent}
\title{godkjent}
\usage{
godkjent(
  profil = c("FHP", "OVP"),
  geoniv = c("K", "F", "B"),
  year = getOption("khfunctions.year")
)
}
\arguments{
\item{profil}{}

\item{geoniv}{}
}
\description{
godkjent
}
