% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{KH_options}
\alias{KH_options}
\title{KH_options}
\usage{
KH_options()
}
\description{
Reads config-khfunctions.yml and sets global options accordingly. Checks if any
option is different from the config-file, and give the option to update.
}
