% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{KHaggreger}
\alias{KHaggreger}
\title{KHaggreger (kb)}
\usage{
KHaggreger(FIL, vals = list(), globs = SettGlobs())
}
\arguments{
\item{FIL}{}

\item{vals}{}

\item{globs}{global parameters, defaults to SettGlobs}

\item{snitt}{}
}
\description{
KHaggreger (kb)
}
