% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{LagAlleFriskvikIndikatorerForKube}
\alias{LagAlleFriskvikIndikatorerForKube}
\title{LagAlleFriskvikIndikatorerForKube (kb)}
\usage{
LagAlleFriskvikIndikatorerForKube(
  KUBEid,
  KUBE,
  FGP,
  modus = NULL,
  batchdate,
  globs = SettGlobs(),
  ...
)
}
\arguments{
\item{KUBEid}{}

\item{KUBE}{}

\item{modus}{}

\item{batchdate}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
LagAlleFriskvikIndikatorerForKube (kb)
}
