% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkube.R
\name{LagFlereKuber}
\alias{LagFlereKuber}
\title{LagFlereKuber
Wrapper aroung LagKUBE, allowing for more than one KUBE to be made simultaneously}
\usage{
LagFlereKuber(KUBEid_ALLE, dumps = list(), alarm = FALSE, ...)
}
\arguments{
\item{...}{Optional, can set versjonert, csvcopy, write arguments if TRUE not wanted}
}
\description{
LagFlereKuber
Wrapper aroung LagKUBE, allowing for more than one KUBE to be made simultaneously
}
