% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{LagFriskvikIndikator}
\alias{LagFriskvikIndikator}
\title{LagFriskvikIndikator (kb)}
\usage{
LagFriskvikIndikator(
  id,
  KUBE = data.table(),
  FGP = list(amin = 0, amax = 120),
  modus = NULL,
  aargang = NULL,
  batchdate = SettKHBatchDate(),
  globs = SettGlobs()
)
}
\arguments{
\item{id}{}

\item{KUBE}{}

\item{FGP}{}

\item{aargang}{}

\item{batchdate}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
LagFriskvikIndikator (kb)
}
