% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkube.R
\name{LagKUBE}
\alias{LagKUBE}
\title{LagKUBE}
\usage{
LagKUBE(
  name,
  write = TRUE,
  alarm = FALSE,
  year = getOption("khfunctions.year"),
  dumps = list(),
  qualcontrol = TRUE,
  ramlimit = NULL
)
}
\arguments{
\item{name}{Name of kube, corresponding to KUBE_NAVN in ACCESS}

\item{write}{should results be written to files, default = TRUE. Set to FALSE for testing (only save to global envir)}

\item{alarm}{if TRUE, plays a sound when done}

\item{year}{year to get valid GEO codes and to produce correct FRISKVIK files, defaults to getOption("khfunctions.year")}

\item{dumps}{list of required dumps, in the format list(dumpname = "format")}

\item{qualcontrol}{perform initial qualcontrol of data}

\item{ramlimit}{max RAM allocation for duckdb. If NULL, default value is 8GB.}
}
\value{
complete data file, publication ready file, and quality control file.
}
\description{
The main function of the production line, producing the files going to FHI Statistikk and public health profiles
}
