% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkube.R
\name{LagKubeDatertCsv}
\alias{LagKubeDatertCsv}
\title{LagKubeDatertCsv
Wrapper around LagKUBE, with default options to save output files}
\usage{
LagKubeDatertCsv(name, write = TRUE, alarm = FALSE, dumps = list())
}
\description{
LagKubeDatertCsv
Wrapper around LagKUBE, with default options to save output files
}
