% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{LagQCKube}
\alias{LagQCKube}
\title{LagQCKube (vl)}
\usage{
LagQCKube(allvis, allvistabs, kube)
}
\arguments{
\item{allvis}{Censored ALLVIs kube}

\item{allvistabs}{Dimensions included in ALLVIS kube}

\item{uprikk}{Uncensored KUBE}

\item{allvisvals}{All columns}
}
\description{
Adds uncensored columns sumTELLER/sumNEVNER/RATE.n to the ALLVISkube
}
