% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_format_filegroups.R
\name{load_and_format_filegroups_to_duckdb}
\alias{load_and_format_filegroups_to_duckdb}
\title{load_and_format_filegroups_to_duckdb}
\usage{
load_and_format_filegroups_to_duckdb(parameters)
}
\arguments{
\item{parameters}{list of parameters generated by get_cubeparameters()}
}
\description{
Read and format filegroups specified in parameters$files into duckdb
Teller file is loaded first, and other files are filtered according to years present in this file
Files are always loaded fresh into duckdb, and overwritten if they already exist.
}
