% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_format_filegroups.R
\name{load_and_format_files}
\alias{load_and_format_files}
\title{load_and_format_filegroups}
\usage{
load_and_format_files(parameters, batchdate, versjonert, globs)
}
\arguments{
\item{parameters}{list of parameters generated by get_cubeparameters()}
}
\value{
updated parameter list after loading and formatting files
}
\description{
Loads filegroups as specified in parameters$files
Adds FileDesign to parameter list after loading files

Teller file is loaded first, and other files are filtered according to years present in this file
}
