% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_and_folders.R
\name{make_godkjent_folder}
\alias{make_godkjent_folder}
\title{make_godkjent_folder}
\usage{
make_godkjent_folder(
  profil = c("FHP", "OVP"),
  geoniv = c("K", "F", "B"),
  year = getOption("khfunctions.year")
)
}
\description{
Generate a new folder with approved friskvik indicator files
}
