% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_from_file.R
\name{make_table_from_original_file}
\alias{make_table_from_original_file}
\title{make_table_from_original_file}
\usage{
make_table_from_original_file(file_number, codebooklog, parameters)
}
\description{
Read and formats an original file before stacking
}
