% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{merge_teller_nevner}
\alias{merge_teller_nevner}
\title{merge_teller_nevner}
\usage{
merge_teller_nevner(parameters, standardfiles = FALSE, design = NULL)
}
\arguments{
\item{parameters}{parameters generated with get_cubeparameters()}

\item{standardfiles}{Should standard teller and nevner files be used? default = FALSE}

\item{design}{Design list}
}
\description{
Merge TELLER/NEVNER files, create initial KUBE and update the empty data.table by reference.
Return the initial design, to be used later.
}
