% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{ModifiserDesign}
\alias{ModifiserDesign}
\title{ModifiserDesign (kb)}
\usage{
ModifiserDesign(Nytt, Org = list(), globs = SettGlobs())
}
\arguments{
\item{Nytt}{}

\item{Org}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
ModifiserDesign (kb)
}
