% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{OmkodFilFraPart}
\alias{OmkodFilFraPart}
\title{OmkodFilFraPart (kb)}
\usage{
OmkodFilFraPart(
  Fil,
  Part,
  FGP = list(amin = 0, amax = 120),
  globs = SettGlobs()
)
}
\arguments{
\item{Fil}{}

\item{Part}{}

\item{FGP}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
OmkodFilFraPart (kb)
}
