% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{organize_file_for_moving_average}
\alias{organize_file_for_moving_average}
\title{organize_file_for_moving_average}
\usage{
organize_file_for_moving_average(dt)
}
\description{
Make sure dt is arranged according to AARl and AARh last
}
