% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_format_filegroups.R
\name{read_filegroup}
\alias{read_filegroup}
\title{read_filegroup}
\usage{
read_filegroup(filegroup)
}
\value{
data.table
}
\description{
Reads original filegroup. Will always read from "PARQUET" folder first.
If filegroup does not exist it will look in "NYESTE" to read the .rds file, and throw an error if not found.
Can be used alone to read filegroups from PARQUET/NYESTE by name.
}
