% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_filegroup_columns.R
\name{recode_columns_with_codebook}
\alias{recode_columns_with_codebook}
\title{recode_columns_with_codebook}
\usage{
recode_columns_with_codebook(
  dt,
  filedescription,
  parameters,
  codebooklog,
  dumps
)
}
\arguments{
\item{dt}{data}

\item{filedescription}{filedescription}

\item{codebooklog}{overal recode log}

\item{dumps}{in case file dumps before or after kodebok are requested.}

\item{codebook}{codebook subset for file}

\item{recode_cols}{columns to recode}
}
\value{
recoded data file (updated by reference)
}
\description{
Recode data according to criteria set in KODEBOK.
Recodings with TYPE = "KB" are directly recoded using do_recode_kb,
and recodings with TYPE = "SUB" are applied as regex using do_recode_regex.
Codebooklog summarises all recodings, and are updated by reference after
each file has been recoded.
}
