% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{save_access_specs}
\alias{save_access_specs}
\title{GetAccessSpecs (vl)}
\usage{
save_access_specs(KUBEid, parameterlist, batchdate, globs = SettGlobs())
}
\arguments{
\item{kuber}{}

\item{tnp}{}

\item{filgrupper}{}

\item{STPNdscr}{}
}
\description{
GetAccessSpecs (vl)
}
