% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{save_kubespec_csv}
\alias{save_kubespec_csv}
\title{kube_spec (ybk)}
\usage{
save_kubespec_csv(
  spec,
  dims = NULL,
  geonaboprikk = NULL,
  geoprikktriangel = NULL
)
}
\description{
Saves ACCESS specs + list of dimensions to be used in STATA censoring
}
