% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_format_filegroups.R
\name{set_lead_value}
\alias{set_lead_value}
\title{set_lead_value

CAN REPLACE YALAGVAL, BUT MUST FIND A WAY OF USING IT IN ACCESS
MAKE NEW COLUMN "LAGVAL"}
\usage{
set_lead_value(file, yearlag = -1, vals = NULL)
}
\arguments{
\item{file}{}

\item{yearlag}{}

\item{vals}{}
}
\description{
set_lead_value

CAN REPLACE YALAGVAL, BUT MUST FIND A WAY OF USING IT IN ACCESS
MAKE NEW COLUMN "LAGVAL"
}
