% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeparameters.R
\name{set_predictionfilter}
\alias{set_predictionfilter}
\title{set_predictionfilter (kb)}
\usage{
set_predictionfilter(parameters, globs)
}
\arguments{
\item{globs}{global parameters, defaults to SettGlobs}

\item{REFVERDI}{Corresponds to ACCESS::KUBER::REFVERDI}
}
\description{
Helper function for \code{get_cubeparameters()}
}
