% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{set_rectangularized_cube_design}
\alias{set_rectangularized_cube_design}
\title{do_rectangularize_cube}
\usage{
set_rectangularized_cube_design(colnames, design, globs = SettGlobs())
}
\arguments{
\item{colnames}{}

\item{design}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
rectangularizes cube based on the given design
}
