% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{set_teller_nevner_names}
\alias{set_teller_nevner_names}
\title{set_teller_nevner_names}
\usage{
set_teller_nevner_names(file, TNPparameters)
}
\description{
Sets name of teller and nevner column to TELLER and NEVNER by reference
}
