% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHglobs.R
\name{SettDefDesignKH}
\alias{SettDefDesignKH}
\title{SettDefDesignKH}
\usage{
SettDefDesignKH(deler)
}
\arguments{
\item{deler}{}
}
\description{
Setter standard designegenskaper, slik som delenes kolonnenavn og status i omkoding
Se tabell KH_DELER
}
\author{
Kåre Bævre
}
