% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{SettFilterDesign}
\alias{SettFilterDesign}
\title{SettFilterDesign (kb)}
\usage{
SettFilterDesign(
  KUBEdscr,
  OrgParts = list(),
  bruk0 = TRUE,
  FGP = list(amin = 0, amax = 120),
  globs = SettGlobs()
)
}
\arguments{
\item{KUBEdscr}{}

\item{OrgParts}{}

\item{bruk0}{}

\item{FGP}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
SettFilterDesign (kb)
}
