% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{SettKHBatchDate}
\alias{SettKHBatchDate}
\title{is_kh_debug (ybk)}
\usage{
SettKHBatchDate()
}
\description{
Debugging all other functions in the project
To use, set show_functions or show_arguments = TRUE
SettKHBatchDate
}
\details{
Used to set date tag in file names
}
