% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHglobs.R
\name{SettKodeBokGlob}
\alias{SettKodeBokGlob}
\title{SettKodeBokGlob}
\usage{
SettKodeBokGlob(globs)
}
\arguments{
\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
SettKodeBokGlob
}
\author{
Kåre Bævre
}
