% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHglobs.R
\name{SettLegitimeKoder}
\alias{SettLegitimeKoder}
\title{SettLegitimeKoder}
\usage{
SettLegitimeKoder(globs)
}
\arguments{
\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
SettLegitimeKoder
}
\author{
Kåre Bævre
}
