% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{SettNaboAnoSpec}
\alias{SettNaboAnoSpec}
\title{SettNaboAnoSpec (kb)}
\usage{
SettNaboAnoSpec(
  ovkatspec,
  FGP = list(amin = 0, amax = 120),
  globs = SettGlobs()
)
}
\arguments{
\item{ovkatspec}{}

\item{FGP}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
SettNaboAnoSpec (kb)
}
