% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R, R/find_redesign.R
\name{SettPartDekk}
\alias{SettPartDekk}
\title{SettPartDekk (kb)}
\usage{
SettPartDekk(
  KB,
  del = "",
  har = paste0(del, "_HAR"),
  betcols = character(0),
  parameters
)

SettPartDekk(
  KB,
  del = "",
  har = paste0(del, "_HAR"),
  betcols = character(0),
  parameters
)
}
\arguments{
\item{KB}{}

\item{del}{}

\item{har}{}

\item{betcols}{}

\item{parameters}{global parameters}

\item{IntervallHull}{}
}
\description{
SettPartDekk (kb)

SettPartDekk (kb)
}
