% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{SkrivKBLogg}
\alias{SkrivKBLogg}
\title{SkrivKBLogg (kb)}
\usage{
SkrivKBLogg(
  KB,
  type,
  filbesk,
  gruppe,
  batchdate = SettKHBatchDate(),
  globs = get_global_parameters()
)
}
\arguments{
\item{KB}{}

\item{type}{}

\item{filbesk}{}

\item{gruppe}{}

\item{batchdate}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
SkrivKBLogg (kb)
}
