% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{uselocal}
\alias{uselocal}
\title{uselocal
for development, sources all files in the R folder}
\usage{
uselocal(test = FALSE, debug = FALSE)
}
\description{
uselocal
for development, sources all files in the R folder
}
