% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_cube_output}
\alias{write_cube_output}
\title{write_cube_output}
\usage{
write_cube_output(
  outputlist,
  name,
  batchdate,
  versjonert,
  csvcopy,
  geonaboprikk
)
}
\arguments{
\item{outputlist}{list of output to write}

\item{name}{KUBEid}

\item{batchdate}{batchdate}

\item{versjonert}{versjonert}

\item{csvcopy}{csvcopy}

\item{geonaboprikk}{geonaboprikk}
}
\description{
Writes KUBE, ALLVIS, and QC files from lagKUBE

REDUCE NUMBER OF ARGUMENTS TO A LIST OF INPUT-args, which can be added to parameters-"input_args" and passed to other functions
}
