% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{do_aggregate_periods}
\alias{do_aggregate_periods}
\title{do_aggregate_periods}
\usage{
do_aggregate_periods(dt, parameters)
}
\arguments{
\item{dt}{data}

\item{parameters}{cube parameters}
}
\description{
Aggregate from original data to moving averages
}
