% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{do_filter_file}
\alias{do_filter_file}
\title{do_filter_file}
\usage{
do_filter_file(file, design, globs = SettGlobs())
}
\arguments{
\item{file}{}

\item{design}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
do_filter_file
}
