% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_handling.R
\name{do_special_handling}
\alias{do_special_handling}
\title{do_special_handling}
\usage{
do_special_handling(dt, code, batchdate, stata_exe, DTout = TRUE)
}
\arguments{
\item{dt}{data to be processed}

\item{code}{code to be performed, either R or STATA}

\item{batchdate}{batchdate, used to create filepaths for stata processing}

\item{globs}{global parameters}
}
\description{
Do special handling of data files, either in R or STATA.
code must be provided in specific files and specified in STATA to be used at certain points in data processing.
}
