% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_redesign.R
\name{find_redesign}
\alias{find_redesign}
\title{find_redesign}
\usage{
find_redesign(orgdesign, targetdesign, aggregate = character(), parameters)
}
\arguments{
\item{orgdesign}{design of file, must be set with FinnDesign()}

\item{targetdesign}{target design, can be simpler than orgdesign as only uses $Part}

\item{aggregate}{aggregate parts? Used in standardization, default = character()}

\item{parameters}{global parameters}
}
\value{
A list with information to redesign a file into a target design.
}
\description{
WIP: replacement for finnredesign()
}
