% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{FinnDesign}
\alias{FinnDesign}
\title{FinnDesign (kb)}
\usage{
FinnDesign(FIL, FGP = list(amin = 0, amax = 120), globs = SettGlobs())
}
\arguments{
\item{FIL}{}

\item{FGP}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
Brukes i lagkube og lagfilgruppe
}
