% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{FinnFilGruppeFraKoblid}
\alias{FinnFilGruppeFraKoblid}
\title{FinnFilGruppeFraKoblid (kb)}
\usage{
FinnFilGruppeFraKoblid(koblid, globs = SettGlobs())
}
\arguments{
\item{koblid}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
FinnFilGruppeFraKoblid (kb)
}
