% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{FinnRedesign}
\alias{FinnRedesign}
\title{FinnRedesign (kb)}
\usage{
FinnRedesign(
  fradesign,
  tildesign,
  SkalAggregeresOpp = character(),
  globs = SettGlobs()
)
}
\arguments{
\item{fradesign}{}

\item{tildesign}{}

\item{SkalAggregeresOpp}{skal noen deler evt aggregeres opp?}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
FinnRedesign (kb)
}
