% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friskvik.R
\name{generate_specific_friskvik_indicators}
\alias{generate_specific_friskvik_indicators}
\title{generate_specific_friskvik_indicators}
\usage{
generate_specific_friskvik_indicators(
  cubename = NULL,
  friskvik_id = NULL,
  year = getOption("khfunctions.year")
)
}
\arguments{
\item{cubename}{name of cube}

\item{friskvik_id}{optional, specify the ID of the indicators you want to generate}

\item{year}{year, defaults to getOption("khfunctions.year")}
}
\description{
generate_specific_friskvik_indicators
}
\examples{
# implement_specific_friskvik_indicators(cubename = "UTDN", friskvik_id = c(1719, 1771), year = 2025) # Specific indicators and year
# implement_specific_friskvik_indicators(cubename = "UTDN") # All indicators, default production year
}
