% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{get_initialdesign}
\alias{get_initialdesign}
\title{get_initialdesign}
\usage{
get_initialdesign(design, tellerfildesign, nevnerfildesign, globs)
}
\arguments{
\item{design}{}

\item{tellerfildesign}{}

\item{nevnerfildesign}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
get_initialdesign
}
