% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{handle_udekk}
\alias{handle_udekk}
\title{handle_udekk (ybk)}
\usage{
handle_udekk(FULL, namesFULL, TempFile)
}
\arguments{
\item{FULL}{}

\item{namesFULL}{}

\item{TempFile}{}
}
\description{
handle_udekk (ybk)
}
