% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khfunctions-package.R
\docType{package}
\name{khfunctions-package}
\alias{khfunctions}
\alias{khfunctions-package}
\title{khfunctions: Data processing for FHP/OVP}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Vegard Lysne \email{vegard.lysne@helsedir.no}

}
\keyword{internal}
