% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHfilgruppe.R
\name{LagFilgruppe}
\alias{LagFilgruppe}
\title{LagFilgruppe}
\usage{
LagFilgruppe(gruppe, versjonert = TRUE, write = TRUE, dumps = list())
}
\arguments{
\item{gruppe}{name of filegroup}

\item{versjonert}{save file with date tag? default = TRUE}

\item{write}{save output files? default = TRUE}

\item{dumps}{list of intermediate files to save, used for debugging and development.}
}
\description{
Loops over all original files, creates a table and append to the complete file group
}
