% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHfilgruppe.R
\name{LagFilgruppe}
\alias{LagFilgruppe}
\title{LagFilgruppe}
\usage{
LagFilgruppe(name, write = TRUE, dumps = list(), qualcontrol = TRUE)
}
\arguments{
\item{write}{save output files? default = TRUE}

\item{dumps}{list of intermediate files to save, used for debugging and development.}

\item{gruppe}{name of filegroup}
}
\description{
Loops over all original files, creates a table and append to the complete file group
}
