% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkube.R
\name{LagKUBE}
\alias{LagKUBE}
\title{LagKUBE}
\usage{
LagKUBE(
  KUBEid,
  versjonert = FALSE,
  csvcopy = FALSE,
  dumps = list(),
  write = FALSE,
  alarm = FALSE,
  geonaboprikk = TRUE,
  ...
)
}
\arguments{
\item{KUBEid}{Name of kube, corresponding to KUBE_NAVN in ACCESS}

\item{versjonert}{}

\item{csvcopy}{Save a CSV-copy?}

\item{dumps}{list of required dumps}

\item{write}{should results be written to files, default = TRUE. Set to FALSE for testing (only save to global envir)}

\item{alarm}{if TRUE, plays a sound when done}

\item{geonaboprikk}{should the file be secondary censored on geographical codes? default = TRUE}

\item{...}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\value{
complete data file, publication ready file, and quality control file.
}
\description{
The main function of the production line, producing the files going to FHI Statistikk and public health profiles
}
