% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkube.R
\name{LagKUBE}
\alias{LagKUBE}
\title{LagKUBE}
\usage{
LagKUBE(
  name,
  write = TRUE,
  alarm = FALSE,
  geonaboprikk = TRUE,
  year = getOption("khfunctions.year"),
  dumps = list(),
  removebuffer = TRUE,
  qualcontrol = FALSE
)
}
\arguments{
\item{name}{Name of kube, corresponding to KUBE_NAVN in ACCESS}

\item{write}{should results be written to files, default = TRUE. Set to FALSE for testing (only save to global envir)}

\item{alarm}{if TRUE, plays a sound when done}

\item{geonaboprikk}{should the file be secondary censored on geographical codes? default = TRUE}

\item{year}{year to get valid GEO codes and to produce correct FRISKVIK files, defaults to getOption("khfunctions.year")}

\item{dumps}{list of required dumps, in the format list(dumpname = "format")}

\item{removebuffer}{should original files in the buffer be removed when no longer needed to free memory?}

\item{qualcontrol}{perform initial qualcontrol of data (default = FALSE for now)}
}
\value{
complete data file, publication ready file, and quality control file.
}
\description{
The main function of the production line, producing the files going to FHI Statistikk and public health profiles
}
