% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkube.R
\name{LagKubeDatertCsv}
\alias{LagKubeDatertCsv}
\title{LagKubeDatertCsv
Wrapper around LagKUBE, with default options to save output files}
\usage{
LagKubeDatertCsv(
  cube_name,
  write = TRUE,
  alarm = FALSE,
  geonaboprikk = TRUE,
  dumps = list()
)
}
\description{
LagKubeDatertCsv
Wrapper around LagKUBE, with default options to save output files
}
