% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkube.R
\name{LagKubeDatertCsv}
\alias{LagKubeDatertCsv}
\title{LagKubeDatertCsv
Wrapper around LagKUBE, with default options to save output files}
\usage{
LagKubeDatertCsv(
  KUBEID,
  dumps = list(),
  versjonert = TRUE,
  csvcopy = TRUE,
  write = TRUE,
  alarm = FALSE
)
}
\description{
LagKubeDatertCsv
Wrapper around LagKUBE, with default options to save output files
}
