% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{merge_teller_nevner}
\alias{merge_teller_nevner}
\title{merge_teller_nevner}
\usage{
merge_teller_nevner(
  parameters,
  standardfiles = FALSE,
  design = NULL,
  globs = SettGlobs()
)
}
\arguments{
\item{parameters}{parameters generated with get_cubeparameters()}

\item{standardfiles}{Should standard teller and nevner files be used? default = FALSE}

\item{design}{Design list}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
merge_teller_nevner
}
