% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_original_file.R
\name{read_original_file}
\alias{read_original_file}
\title{read_original_file}
\usage{
read_original_file(filedescription, parameters, dumps = list())
}
\arguments{
\item{filedescription}{Row from parameters$read_parameters corresponding to the current file}

\item{parameters}{filegroup parameters, set by \code{get_filegroup_parameters()}}

\item{dumps}{any file dumps requested}
}
\value{
formatted original file, ready for further processing
}
\description{
Reads the original file, with specific methods to read XLS/XLSX/CSV/SPSS-files,
all producing data.table for further processing. Excel/CSV-files are further formatted
according to parameters set inn INNLESARGS (or default values).
\itemize{
\item Sets manual header as specified in MANHEADER
\item Applies special handling if code provided to RSYNT1
}
}
