% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_teller_nevner.R
\name{report_removed_codes}
\alias{report_removed_codes}
\title{get_removed_codes}
\usage{
report_removed_codes(file, cube)
}
\description{
fetches info on geo codes removed during rectangularization
}
