% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{save_filedump_if_requested}
\alias{save_filedump_if_requested}
\title{save_filedump_if_requested}
\usage{
save_filedump_if_requested(
  dumpname = c("RSYNT1pre", "RSYNT1post", "RESHAPEpre", "RESHAPEpost", "RSYNT2pre",
    "RSYNT2post", "KODEBOKpre", "KODEBOKpost", "RSYNT_PRE_FGLAGRINGpre",
    "RSYNT_PRE_FGLAGRINGpost", "MOVAVpre", "MOVAVpost", "SLUTTREDIGERpre",
    "SLUTTREDIGERpost", "PRIKKpre", "PRIKKpost", "RSYNT_POSTPROSESSpre",
    "RSYNT_POSTPROSESSpost"),
  dt,
  parameters,
  koblid = NULL
)
}
\arguments{
\item{dumpname}{name of requested filedump}

\item{dt}{data to be stored}

\item{parameters}{global parameters}

\item{koblid}{used when requesting file dumps during processing of original files, default = NULL}
}
\description{
Saves a .csv, .rds, or .dta file at specific points during data processing
All RSYNT points have the possibility to save a filedump with the name rsyntname + pre/post, e.g. "RSYNT_POSTPROSESSpre"

Filegroup processing:
\itemize{
\item RSYNT1pre/post (Rsynt when reading original file)
\item RESHAPEpre/post (Before and after reshaping original file)
\item RSYNT2pre/post (Rsynt during formatting of original file to table)
\item KODEBOKpre/post (before and after recoding with codebook)
\item RSYNT_PRE_FGLAGRINGpre/post (Before and after rsynt point prior to saving output)
}

Cube processing:
\itemize{
\item MOVAVpre/MOVAVpost (before and after aggregating to periods)
\item SLUTTREDIGERpre/post (before and after RSYNT point SLUTTREDIGER, after aggregation and standardization)
\item PRIKKpre/post (before and after censoring)
\item RSYNT_POSTPROSESSpre/post (Before and after RSYNT_POSTPROSESS)
}
}
\examples{
# LagKUBE("KUBENAVN", dumps = list(PRIKKpre = "STATA", PRIKKpost = c("CSV", "STATA", "R"))
# LagFilgruppe("FILGRUPPENAVN", dumps = list(RSYNT1pre = "STATA", KODEBOKpost = c("CSV", "STATA", "R") )
}
