% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeparameters.R
\name{SettKodeBokGlob}
\alias{SettKodeBokGlob}
\title{SettKodeBokGlob}
\usage{
SettKodeBokGlob(parameters)
}
\arguments{
\item{parameters}{global parameters}
}
\description{
SettKodeBokGlob
}
\author{
Kåre Bævre
}
