% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHkubefunctions.R
\name{SettPartDekk}
\alias{SettPartDekk}
\title{SettPartDekk (kb)}
\usage{
SettPartDekk(
  KB,
  del = "",
  har = paste0(del, "_HAR"),
  betcols = character(0),
  IntervallHull = globs$DefDesign$IntervallHull,
  globs = SettGlobs()
)
}
\arguments{
\item{KB}{}

\item{del}{}

\item{har}{}

\item{betcols}{}

\item{IntervallHull}{SettPartDekk(KBD, del = del, har = delH, IntervallHull = IntervallHull, globs = globs)}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
SettPartDekk (kb)
}
