% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHmisc.R
\name{TilFilLogg}
\alias{TilFilLogg}
\title{TilFilLogg (kb)}
\usage{
TilFilLogg(
  koblid,
  felt,
  verdi,
  batchdate = SettKHBatchDate(),
  globs = get_global_parameters()
)
}
\arguments{
\item{koblid}{}

\item{felt}{}

\item{verdi}{}

\item{batchdate}{}

\item{globs}{global parameters, defaults to SettGlobs}
}
\description{
TilFilLogg (kb)
}
